Reports Development
Reports Controller and Manager Developed
DataMock Added for historical data

+++ digitus-trial/Digitus.Trial.Backend.Api.Security/ApiModels/LoginReportResultModel.cs
@@ -4,6 +4,6 @@ namespace Digitus.Trial.Backend.Api.ApiModels
     public class LoginReportResultModel
     {
         public DateTime ReportDate { get; set; }
-        public long AverageLoginDuration { get; set; }
+        public double AverageLoginDuration { get; set; }
     }
 }
